refactor: separate auth from OAuth
