+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/user/service/AuthService.kt
@@ -6,6 +6,4 @@ import com.sexysisters.tojserverv2.domain.user.UserInfo
 interface AuthService {
     fun login(request: UserCommand.LoginRequest): UserInfo.Token
     fun logout(accessToken: String)
-    fun getGoogleLink(): String
-    fun googleLogin(code: String): UserInfo.Token
 }
