+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/user/service/OAuthService.kt
@@ -1,0 +1,8 @@
+package com.sexysisters.tojserverv2.domain.user.service
+
+import com.sexysisters.tojserverv2.domain.user.UserInfo
+
+interface OAuthService {
+    fun getGoogleLink(): String
+    fun googleLogin(code: String): UserInfo.Token
+}
