+++ src/main/kotlin/com/sexysisters/tojserverv2/interfaces/user/AuthApiController.kt
@@ -9,14 +9,12 @@ import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import org.mapstruct.factory.Mappers
 import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import javax.validation.Valid
-import javax.websocket.server.PathParam
 
 @Api(tags = ["Auth 관련 API"])
 @RestController
@@ -39,19 +37,4 @@ class AuthApiController(
     @DeleteMapping
     fun logout(@RequestHeader("Authorization") accessToken: String) =
         authService.logout(accessToken)
-
-    @ApiOperation(value = "구글 OAuth 로그인 링크")
-    @GetMapping("/google")
-    fun queryGoogleAuthLink(): CommonResponse<String> {
-        val link = authService.getGoogleLink()
-        return CommonResponse.success(link)
-    }
-
-    @ApiOperation(value = "Google OAuth callback (client 사용 안함)")
-    @GetMapping("/google/callback")
-    fun googleAuth(@PathParam("code") code: String): CommonResponse<UserResponse.Token> {
-        val userInfo = authService.googleLogin(code)
-        val response: UserResponse.Token = mapper.of(userInfo)
-        return CommonResponse.success(response)
-    }
 }
